Crate boa_temporal

Boa’s boa_temporal crate is intended to serve as an engine agnostic implementation the ECMAScript’s Temporal builtin and algorithm.

About Boa

Boa is an open-source, experimental ECMAScript Engine written in Rust for lexing, parsing and executing ECMAScript/JavaScript. Currently, Boa supports some of the language. More information can be viewed at Boa’s website.

Try out the most recent release with Boa’s live demo playground.

Boa Crates

boa_ast - Boa’s ECMAScript Abstract Syntax Tree.

boa_engine - Boa’s implementation of ECMAScript builtin objects and execution.

boa_gc - Boa’s garbage collector.

boa_interner - Boa’s string interner.

boa_parser - Boa’s lexer and parser.

boa_profiler - Boa’s code profiler.

boa_icu_provider - Boa’s ICU4X data provider.

boa_runtime - Boa’s WebAPI features.

Re-exports

pub use error::TemporalError;

Modules

calendar

Temporal calendar traits and implementations.

date

The PlainDate representation.

datetime

Temporal implementation of DateTime

duration

The Temporal Duration.

error

An error type for Temporal Errors.

fields

TemporalFields native Rust representation.

iso

The ISO module implements the internal ISO field slots.

month_day

MonthDay

options

Temporal Options

parser

Implementation of Iso8601 grammar lexing/parsing

time

Temporal Time Representation.

utils 🔒

Utility equations for Temporal

year_month

YearMonth

zoneddatetime

The ZonedDateTime module.

Structs

TinyAsciiStr

Constants

MS_PER_DAY 🔒

Milliseconds per day constant: 8.64e+7

NS_MAX_INSTANT 🔒

Max Instant nanosecond constant

NS_MIN_INSTANT 🔒

Min Instant nanosecond constant

NS_PER_DAY 🔒

Nanoseconds per day constant: 8.64e+13

Type Aliases

TemporalResult

The Temporal result type