Module boa_temporal::components

The primary date-time components provided by Temporal.

The below components are the main primitives of the Temporal specification:

Date -> PlainDate

DateTime -> PlainDateTime

Time -> PlainTime

Duration -> Duration

Instant -> Instant

MonthDay -> PlainMonthDay

YearMonth -> PlainYearMonth

ZonedDateTime -> ZonedDateTime

The Temporal specification, along with this implementation aims to provide full support for time zones and non-gregorian calendars that are compliant with standards like ISO 8601, RFC 3339, and RFC 5545.

Modules

calendar

This module implements the calendar traits and related components.

date 🔒

This module implements Date and any directly related algorithms.

datetime 🔒

This module implements DateTime any directly related algorithms.

duration 🔒

This module implements Duration along with it’s methods and components.

instant 🔒

An implementation of the Temporal Instant.

month_day 🔒

This module implements MonthDay and any directly related algorithms.

time 🔒

This module implements Time and any directly related algorithms.

tz

This module implements the Temporal TimeZone and components.

year_month 🔒

This module implements YearMonth and any directly related algorithms.

zoneddatetime 🔒

This module implements ZonedDateTime and any directly related algorithms.

Structs

Date

The native Rust implementation of Temporal.PlainDate.

DateTime

The native Rust implementation of Temporal.PlainDateTime

Duration

The native Rust implementation of Temporal.Duration.

Instant

The native Rust implementation of Temporal.Instant

MonthDay

The native Rust implementation of Temporal.PlainMonthDay

Time

The native Rust implementation of Temporal.PlainTime.

YearMonth

The native Rust implementation of Temporal.YearMonth.

ZonedDateTime

The native Rust implementation of Temporal.ZonedDateTime.