Module boa_temporal::components::calendar

This module implements the calendar traits and related components.

The goal of the calendar module of boa_temporal is to provide Temporal compatible calendar implementations.

The implementation will only be of calendar’s prexisting calendars. This library does not come with a pre-existing CustomCalendar (i.e., an object that implements the calendar protocol), but it does aim to provide the necessary tools and API for implementing one.

Enums

CalendarDateLike

The DateLike objects that can be provided to the CalendarProtocol.

CalendarFieldsType

Designate the type of CalendarFields needed

CalendarSlot

The [[Calendar]] field slot of a Temporal Object.

Constants

CALENDAR_PROTOCOL_METHODS

The ECMAScript defined protocol methods

Traits

CalendarProtocol

A trait for implementing a Builtin Calendar’s Calendar Protocol in Rust.

GetCalendarSlot

A trait for retrieving an internal calendar slice.