Struct boa_temporal::components::date::Date

pub struct Date<C: CalendarProtocol> {
    iso: IsoDate,
    calendar: CalendarSlot<C>,
}

The native Rust implementation of Temporal.PlainDate.

Fields
iso: IsoDatecalendar: CalendarSlot<C>
Implementations

impl<C: CalendarProtocol> Date<C>

pub(crate) fn new_unchecked(iso: IsoDate, calendar: CalendarSlot<C>) -> Self

Create a new Date with the date values and calendar slot.

pub(crate) fn move_relative_date( &self, duration: &Duration, context: &mut C::Context ) -> TemporalResult<(Self, f64)>

Returns a new moved date and the days associated with that adjustment

impl<C: CalendarProtocol> Date<C>

pub fn new( year: i32, month: i32, day: i32, calendar: CalendarSlot<C>, overflow: ArithmeticOverflow ) -> TemporalResult<Self>

Creates a new Date while checking for validity.

pub fn from_datetime(dt: &DateTime<C>) -> Self

Creates a Date from a DateTime.

pub const fn iso_year(&self) -> i32

Returns this Date’s ISO year value.

pub const fn iso_month(&self) -> u8

Returns this Date’s ISO month value.

pub const fn iso_day(&self) -> u8

Returns this Date’s ISO day value.

pub const fn iso(&self) -> IsoDate

Returns the Date’s inner IsoDate record.

pub fn calendar(&self) -> &CalendarSlot<C>

Returns a reference to this Date’s calendar slot.

pub fn is_valid(&self) -> bool

3.5.7 IsValidISODate

Checks if the current date is a valid ISODate.

pub fn days_until(&self, other: &Self) -> i32

DaysUntil

Calculates the epoch days between two Dates

impl Date<()>

pub fn year(&self) -> TemporalResult<i32>

Returns the calendar year value.

pub fn month(&self) -> TemporalResult<u8>

Returns the calendar month value.

pub fn month_code(&self) -> TemporalResult<TinyAsciiStr<4>>

Returns the calendar month code value.

pub fn day(&self) -> TemporalResult<u8>

Returns the calendar day value.

pub fn day_of_week(&self) -> TemporalResult<u16>

Returns the calendar day of week value.

pub fn day_of_year(&self) -> TemporalResult<u16>

Returns the calendar day of year value.

pub fn week_of_year(&self) -> TemporalResult<u16>

Returns the calendar week of year value.

pub fn year_of_week(&self) -> TemporalResult<i32>

Returns the calendar year of week value.

pub fn days_in_week(&self) -> TemporalResult<u16>

Returns the calendar days in week value.

pub fn days_in_month(&self) -> TemporalResult<u16>

Returns the calendar days in month value.

pub fn days_in_year(&self) -> TemporalResult<u16>

Returns the calendar days in year value.

pub fn months_in_year(&self) -> TemporalResult<u16>

Returns the calendar months in year value.

pub fn in_leap_year(&self) -> TemporalResult<bool>

Returns returns whether the date in a leap year for the given calendar.

impl<C: CalendarProtocol> Date<C>

pub fn contextual_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<i32>

Returns the calendar year value with provided context.

pub fn contextual_month( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u8>

Returns the calendar month value with provided context.

pub fn contextual_month_code( this: &C::Date, context: &mut C::Context ) -> TemporalResult<TinyAsciiStr<4>>

Returns the calendar month code value with provided context.

pub fn contextual_day( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u8>

Returns the calendar day value with provided context.

pub fn contextual_day_of_week( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar day of week value with provided context.

pub fn contextual_day_of_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar day of year value with provided context.

pub fn contextual_week_of_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar week of year value with provided context.

pub fn contextual_year_of_week( this: &C::Date, context: &mut C::Context ) -> TemporalResult<i32>

Returns the calendar year of week value with provided context.

pub fn contextual_days_in_week( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar days in week value with provided context.

pub fn contextual_days_in_month( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar days in month value with provided context.

pub fn contextual_days_in_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar days in year value with provided context.

pub fn contextual_months_in_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<u16>

Returns the calendar months in year value with provided context.

pub fn contextual_in_leap_year( this: &C::Date, context: &mut C::Context ) -> TemporalResult<bool>

Returns whether the date is in a leap year for the given calendar with provided context.

impl<C: CalendarProtocol> Date<C>

pub fn contextual_add_date( &self, duration: &Duration, overflow: ArithmeticOverflow, context: &mut C::Context ) -> TemporalResult<Self>

Returns the date after adding the given duration to date with a provided context.

Temporal Equivalent: 3.5.13 AddDate ( calendar, plainDate, duration [ , options [ , dateAdd ] ] )

pub fn contextual_difference_date( &self, other: &Self, largest_unit: TemporalUnit, context: &mut C::Context ) -> TemporalResult<Duration>

Returns a duration representing the difference between the dates one and two with a provided context.

Temporal Equivalent: 3.5.6 DifferenceDate ( calendar, one, two, options )