Struct boa_temporal::components::datetime::DateTime

pub struct DateTime<C: CalendarProtocol> {
    iso: IsoDateTime,
    calendar: CalendarSlot<C>,
}

The native Rust implementation of Temporal.PlainDateTime

Fields
iso: IsoDateTimecalendar: CalendarSlot<C>
Implementations

impl<C: CalendarProtocol> DateTime<C>

pub(crate) fn new_unchecked(iso: IsoDateTime, calendar: CalendarSlot<C>) -> Self

Creates a new unchecked DateTime.

fn validate_iso(iso: IsoDate) -> bool

Utility function for validating IsoDates

pub(crate) fn from_instant( instant: &Instant, offset: f64, calendar: CalendarSlot<C> ) -> TemporalResult<Self>

Create a new DateTime from an Instant.

pub(crate) fn iso_date(&self) -> &IsoDate

Returns the inner IsoDate value.

impl<C: CalendarProtocol> DateTime<C>

pub fn new( year: i32, month: i32, day: i32, hour: i32, minute: i32, second: i32, millisecond: i32, microsecond: i32, nanosecond: i32, calendar: CalendarSlot<C> ) -> TemporalResult<Self>

Creates a new validated DateTime.

pub fn validate<T: IsoDateSlots>(target: &T) -> bool

Validates whether ISO date slots are within iso limits at noon.

pub const fn iso_year(&self) -> i32

Returns this Date’s ISO year value.

pub const fn iso_month(&self) -> u8

Returns this Date’s ISO month value.

pub const fn iso_day(&self) -> u8

Returns this Date’s ISO day value.

pub fn hour(&self) -> u8

Returns the hour value

pub fn minute(&self) -> u8

Returns the minute value

pub fn second(&self) -> u8

Returns the second value

pub fn millisecond(&self) -> u16

Returns the millisecond value

pub fn microsecond(&self) -> u16

Returns the microsecond value

pub fn nanosecond(&self) -> u16

Returns the nanosecond value

pub fn calendar(&self) -> &CalendarSlot<C>

Returns the Calendar value.

impl<C: CalendarProtocol> DateTime<C>

pub fn contextual_year(&self, context: &mut dyn Any) -> TemporalResult<i32>

Returns the calendar year value with provided context.

pub fn year(&self) -> TemporalResult<i32>

Returns the calendar year value.

pub fn contextual_month(&self, context: &mut dyn Any) -> TemporalResult<u8>

Returns the calendar month value with provided context.

pub fn month(&self) -> TemporalResult<u8>

Returns the calendar month value.

pub fn contextual_month_code( &self, context: &mut dyn Any ) -> TemporalResult<TinyAsciiStr<4>>

Returns the calendar month code value with provided context.

pub fn month_code(&self) -> TemporalResult<TinyAsciiStr<4>>

Returns the calendar month code value.

pub fn contextual_day(&self, context: &mut dyn Any) -> TemporalResult<u8>

Returns the calendar day value with provided context.

pub fn day(&self) -> TemporalResult<u8>

Returns the calendar day value.

pub fn contextual_day_of_week( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar day of week value with provided context.

pub fn day_of_week(&self) -> TemporalResult<u16>

Returns the calendar day of week value.

pub fn contextual_day_of_year( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar day of year value with provided context.

pub fn day_of_year(&self) -> TemporalResult<u16>

Returns the calendar day of year value.

pub fn contextual_week_of_year( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar week of year value with provided context.

pub fn week_of_year(&self) -> TemporalResult<u16>

Returns the calendar week of year value.

pub fn contextual_year_of_week( &self, context: &mut dyn Any ) -> TemporalResult<i32>

Returns the calendar year of week value with provided context.

pub fn year_of_week(&self) -> TemporalResult<i32>

Returns the calendar year of week value.

pub fn contextual_days_in_week( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar days in week value with provided context.

pub fn days_in_week(&self) -> TemporalResult<u16>

Returns the calendar days in week value.

pub fn contextual_days_in_month( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar days in month value with provided context.

pub fn days_in_month(&self) -> TemporalResult<u16>

Returns the calendar days in month value.

pub fn contextual_days_in_year( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar days in year value with provided context.

pub fn days_in_year(&self) -> TemporalResult<u16>

Returns the calendar days in year value.

pub fn contextual_months_in_year( &self, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the calendar months in year value with provided context.

pub fn months_in_year(&self) -> TemporalResult<u16>

Returns the calendar months in year value.

pub fn contextual_in_leap_year( &self, context: &mut dyn Any ) -> TemporalResult<bool>

Returns whether the date is in a leap year for the given calendar with provided context.

pub fn in_leap_year(&self) -> TemporalResult<bool>

Returns returns whether the date in a leap year for the given calendar.