Struct boa_temporal::components::duration::date::DateDuration

pub struct DateDuration {
    pub(crate) years: f64,
    pub(crate) months: f64,
    pub(crate) weeks: f64,
    pub(crate) days: f64,
}

DateDuration represents the date duration record of the Duration.

These fields are laid out in the Temporal Proposal as 64-bit floating point numbers.

Fields
years: f64months: f64weeks: f64days: f64
Implementations

impl DateDuration

pub(crate) const fn new_unchecked( years: f64, months: f64, weeks: f64, days: f64 ) -> Self

Creates a new, non-validated DateDuration.

impl DateDuration

pub fn new( years: f64, months: f64, weeks: f64, days: f64 ) -> TemporalResult<Self>

Creates a new DateDuration with provided values.

pub const fn partial() -> Self

Returns a PartialDateDuration with all fields set to NaN.

pub fn from_partial(partial: &DateDuration) -> Self

Creates a DateDuration from a provided partial DateDuration.

pub fn abs(&self) -> Self

Returns a new DateDuration representing the absolute value of the current.

pub const fn years(&self) -> f64

Returns the [[years]] value.

pub const fn months(&self) -> f64

Returns the [[months]] value.

pub const fn weeks(&self) -> f64

Returns the [[weeks]] value.

pub const fn days(&self) -> f64

Returns the [[days]] value.

pub fn iter(&self) -> DateIter<'_>

Returns the iterator for DateDuration

impl DateDuration

pub fn round<C: CalendarProtocol, Z: TzProtocol>( &self, additional_time: Option<TimeDuration>, increment: f64, unit: TemporalUnit, rounding_mode: TemporalRoundingMode, relative_targets: (Option<&Date<C>>, Option<&ZonedDateTime<C, Z>>, Option<&DateTime<C>>), context: &mut C::Context ) -> TemporalResult<(Self, f64)>

Rounds the current DateDuration returning a tuple of the rounded DateDuration and the total value of the smallest unit prior to rounding.