Struct boa_temporal::components::duration::time::TimeDuration

pub struct TimeDuration {
    pub(crate) hours: f64,
    pub(crate) minutes: f64,
    pub(crate) seconds: f64,
    pub(crate) milliseconds: f64,
    pub(crate) microseconds: f64,
    pub(crate) nanoseconds: f64,
}

TimeDuration represents the Time Duration record of the Duration.

These fields are laid out in the Temporal Proposal as 64-bit floating point numbers.

Fields
hours: f64minutes: f64seconds: f64milliseconds: f64microseconds: f64nanoseconds: f64
Implementations

impl TimeDuration

pub(crate) const fn new_unchecked( hours: f64, minutes: f64, seconds: f64, milliseconds: f64, microseconds: f64, nanoseconds: f64 ) -> Self

Creates a new TimeDuration.

pub(crate) fn as_nanos(&self) -> f64

Returns the current TimeDuration as nanoseconds.

pub(crate) fn balance_possibly_infinite_time_duration( days: f64, hours: f64, minutes: f64, seconds: f64, milliseconds: f64, microseconds: f64, nanoseconds: f64, largest_unit: TemporalUnit ) -> TemporalResult<(f64, Option<Self>)>

Abstract Operation 7.5.18 BalancePossiblyInfiniteDuration ( days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, largestUnit )

This function will balance the current TimeDuration. It returns the balanced day and TimeDuration value.

impl TimeDuration

pub fn new( hours: f64, minutes: f64, seconds: f64, milliseconds: f64, microseconds: f64, nanoseconds: f64 ) -> TemporalResult<Self>

Creates a new validated TimeDuration.

pub const fn partial() -> Self

Creates a partial TimeDuration with all values set to NaN.

pub fn from_partial(partial: &TimeDuration) -> Self

Creates a TimeDuration from a provided partial TimeDuration.

pub fn abs(&self) -> Self

Returns a new TimeDuration representing the absolute value of the current.

pub fn neg(&self) -> Self

Returns a negated TimeDuration.

pub fn balance( &self, days: f64, largest_unit: TemporalUnit ) -> TemporalResult<(f64, Self)>

Balances a TimeDuration given a day value and the largest unit. balance will return the balanced day and TimeDuration.

Errors:

Will error if provided duration is invalid

pub fn is_within_range(&self) -> bool

Utility function for returning if values in a valid range.

pub const fn hours(&self) -> f64

Returns the [[hours]] value.

pub const fn minutes(&self) -> f64

Returns the [[minutes]] value.

pub const fn seconds(&self) -> f64

Returns the [[seconds]] value.

pub const fn milliseconds(&self) -> f64

Returns the [[milliseconds]] value.

pub const fn microseconds(&self) -> f64

Returns the [[microseconds]] value.

pub const fn nanoseconds(&self) -> f64

Returns the [[nanoseconds]] value.

pub fn iter(&self) -> TimeIter<'_>

Returns the TimeDuration’s iterator.

impl TimeDuration

pub fn round( &self, increment: f64, unit: TemporalUnit, rounding_mode: TemporalRoundingMode ) -> TemporalResult<(Self, f64)>

Rounds the current TimeDuration given a rounding increment, unit and rounding mode. round will return a tuple of the rounded TimeDuration and the total value of the smallest unit prior to rounding.