Struct boa_temporal::components::Duration

pub struct Duration {
    date: DateDuration,
    time: TimeDuration,
}

The native Rust implementation of Temporal.Duration.

Duration is made up of a DateDuration and TimeDuration as primarily defined by Abtract Operation 7.5.1-5.

Fields
date: DateDurationtime: TimeDuration
Implementations

impl Duration

pub(crate) const fn new_unchecked( date: DateDuration, time: TimeDuration ) -> Self

Creates a new Duration from a DateDuration and TimeDuration.

pub(crate) fn one_year(year_value: f64) -> Self

Utility function to create a year duration.

pub(crate) fn one_month(month_value: f64) -> Self

Utility function to create a month duration.

pub(crate) fn one_week(week_value: f64) -> Self

Utility function to create a week duration.

impl Duration

pub fn new( years: f64, months: f64, weeks: f64, days: f64, hours: f64, minutes: f64, seconds: f64, milliseconds: f64, microseconds: f64, nanoseconds: f64 ) -> TemporalResult<Self>

Creates a new validated Duration.

pub const fn partial() -> Self

Creates a partial Duration with all fields set to NaN.

pub fn from_date_duration(date: DateDuration) -> Self

Creates a Duration from only a DateDuration.

pub fn from_day_and_time(day: f64, time: TimeDuration) -> Self

Creates a Duration from a provided a day and a TimeDuration.

Note: TimeDuration records can store a day value to deal with overflow.

pub fn from_partial(partial: &Duration) -> TemporalResult<Self>

Creates a new valid Duration from a partial Duration.

pub fn is_time_within_range(&self) -> bool

Return if the Durations values are within their valid ranges.

pub fn is_date_duration(&self) -> bool

Returns whether Duration’s DateDuration isn’t empty and is therefore a DateDuration or Duration.

pub fn is_time_duration(&self) -> bool

Returns whether Duration’s DateDuration is empty and is therefore a TimeDuration.

impl Duration

pub fn time(&self) -> &TimeDuration

Returns a reference to the inner TimeDuration

pub fn date(&self) -> &DateDuration

Returns a reference to the inner DateDuration

pub fn set_time_duration(&mut self, time: TimeDuration)

Set this DurationRecord’s TimeDuration.

pub fn set_years(&mut self, y: f64)

Set the value for years.

pub const fn years(&self) -> f64

Returns the years field of duration.

pub fn set_months(&mut self, mo: f64)

Set the value for months.

pub const fn months(&self) -> f64

Returns the months field of duration.

pub fn set_weeks(&mut self, w: f64)

Set the value for weeks.

pub const fn weeks(&self) -> f64

Returns the weeks field of duration.

pub fn set_days(&mut self, d: f64)

Set the value for days.

pub const fn days(&self) -> f64

Returns the weeks field of duration.

pub fn set_hours(&mut self, h: f64)

Set the value for hours.

pub const fn hours(&self) -> f64

Returns the hours field of duration.

pub fn set_minutes(&mut self, m: f64)

Set the value for minutes.

pub const fn minutes(&self) -> f64

Returns the hours field of duration.

pub fn set_seconds(&mut self, s: f64)

Set the value for seconds.

pub const fn seconds(&self) -> f64

Returns the seconds field of duration.

pub fn set_milliseconds(&mut self, ms: f64)

Set the value for milliseconds.

pub const fn milliseconds(&self) -> f64

Returns the hours field of duration.

pub fn set_microseconds(&mut self, mis: f64)

Set the value for microseconds.

pub const fn microseconds(&self) -> f64

Returns the microseconds field of duration.

pub fn set_nanoseconds(&mut self, ns: f64)

Set the value for nanoseconds.

pub const fn nanoseconds(&self) -> f64

Returns the nanoseconds field of duration.

pub fn iter(&self) -> DurationIter<'_>

Returns Duration’s iterator

impl Duration

pub(crate) fn unbalance_duration_relative<C: CalendarProtocol>( &self, largest_unit: TemporalUnit, plain_relative_to: Option<&Date<C>>, context: &mut dyn Any ) -> TemporalResult<DateDuration>

7.5.21 UnbalanceDateDurationRelative ( years, months, weeks, days, largestUnit, plainRelativeTo )

pub fn balance_date_duration_relative<C: CalendarProtocol>( &self, largest_unit: TemporalUnit, plain_relative_to: Option<&Date<C>>, context: &mut dyn Any ) -> TemporalResult<DateDuration>

BalanceDateDurationRelative

pub fn round_duration<C: CalendarProtocol, Z: TzProtocol>( &self, increment: f64, unit: TemporalUnit, rounding_mode: TemporalRoundingMode, relative_targets: (Option<&Date<C>>, Option<&ZonedDateTime<C, Z>>, Option<&DateTime<C>>), context: &mut dyn Any ) -> TemporalResult<(Self, f64)>

Abstract Operation 7.5.26 RoundDuration ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, increment, unit, roundingMode [ , plainRelativeTo [, zonedRelativeTo [, precalculatedDateTime]]] )

impl Duration

pub fn abs(&self) -> Self

Returns the absolute value of Duration.

pub fn duration_sign(&self) -> i32

7.5.10 DurationSign ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds )

Determines the sign for the current self.

pub fn default_temporal_largest_unit(&self) -> TemporalUnit

7.5.12 DefaultTemporalLargestUnit ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds )

pub fn balance_time_duration( &self, unit: TemporalUnit ) -> TemporalResult<(f64, TimeDuration)>

Calls TimeDuration’s balance method on the current Duration.