Module boa_temporal::iso

This module implements the internal ISO field slots.

The three main types of slots are:

IsoDateTime

IsoDate

IsoTime

An IsoDate represents the [[ISOYear]], [[ISOMonth]], and [[ISODay]] internal slots.

An IsoTime represents the [[ISOHour]], [[ISOMinute]], [[ISOsecond]], [[ISOmillisecond]], [[ISOmicrosecond]], and [[ISOnanosecond]] internal slots.

An IsoDateTime has the internal slots of both an IsoDate and IsoTime.

Structs

IsoDate

IsoDate serves as a record for the [[ISOYear]], [[ISOMonth]], and [[ISODay]] internal fields.

IsoDateTime

IsoDateTime is the record of the IsoDate and IsoTime internal slots.

IsoTime

An IsoTime record that contains Temporal’s time slots.

Traits

IsoDateSlots

A trait for accessing the IsoDate across the various Temporal objects

Functions

div_mod 🔒

is_valid_date 🔒

is_valid_time 🔒

iso_date_to_epoch_days 🔒

Returns the Epoch days based off the given year, month, and day.

iso_dt_within_valid_limits 🔒

Utility function to determine if a DateTime’s components create a DateTime within valid limits

utc_epoch_nanos 🔒

Utility function to convert a IsoDate and IsoTime values into epoch nanoseconds