Module boa_temporal::options

Native implementation of the Temporal options.

Temporal has various instances where user’s can define options for how an operation may be completed.

Structs

ParseArithmeticOverflowError

A parsing error for ArithemeticOverflow

ParseDurationOverflowError

A parsing error for DurationOverflow.

ParseInstantDisambiguationError

A parsing error on InstantDisambiguation options.

ParseOffsetDisambiguationError

A parsing error for OffsetDisambiguation parsing.

ParseTemporalUnitError

A parsing error for TemporalUnit

Enums

ArithmeticOverflow

ArithmeticOverflow can also be used as an assignment overflow and consists of the “constrain” and “reject” options.

DurationOverflow

Duration overflow options.

InstantDisambiguation

The disambiguation options for an instant.

OffsetDisambiguation

Offset disambiguation options.

TemporalRoundingMode

Declares the specified RoundingMode for the operation.

TemporalUnit

The relevant unit that should be used for the operation that this option is provided as a value.

TemporalUnsignedRoundingMode

The UnsignedRoundingMode