Module boa_temporal::parser::annotations

Structs

AnnotationSet 🔒

Strictly a Parsing Intermediary for the checking the common annotation backing.

KeyValueAnnotation 🔒

A KeyValueAnnotation Parse Node.

RecognizedAnnotations 🔒

An internal crate type to house any recognized annotations that are found.

Functions

parse_annotation_key 🔒

Parse an AnnotationKey.

parse_annotation_set 🔒

Parse a TimeZoneAnnotation Annotations set

parse_annotation_value 🔒

Parse an AnnotationValue.

parse_annotations 🔒

Parse any number of KeyValueAnnotations

parse_kv_annotation 🔒

Parse an annotation with an AnnotationKey=AnnotationValue pair.