Struct boa_temporal::parser::Cursor

pub struct Cursor {
    pos: u32,
    source: Vec<char>,
}

Cursor is a small cursor implementation for parsing Iso8601 grammar.

Fields
pos: u32source: Vec<char>
Implementations

impl Cursor

pub fn new(source: &str) -> Self

Create a new cursor from a source String value.

fn slice(&self, start: u32, end: u32) -> String

Returns a string value from a slice of the cursor.

const fn pos(&self) -> u32

Get current position

fn peek(&self) -> Option<char>

Peek the value at next position (current + 1).

fn peek_n(&self, n: u32) -> Option<char>

Peek the value at n len from current.

fn check<F>(&self, f: F) -> Option<bool>
where F: FnOnce(char) -> bool,

Runs the provided check on the current position.

fn check_or<F>(&self, default: bool, f: F) -> bool
where F: FnOnce(char) -> bool,

Runs the provided check on current position returns the default value if None.

fn next(&mut self) -> Option<char>

Returns Cursor’s current char and advances to the next position.

fn expect_next(&mut self) -> char

Utility method that returns next charactor unwrapped char

Panics

This will panic if the next value has not been confirmed to exist.

fn abrupt_next(&mut self) -> TemporalResult<char>

A utility next method that returns an AbruptEnd error if invalid.

fn advance(&mut self)

Advances the cursor’s position by 1.

fn advance_if(&mut self, condition: bool)

Utility function to advance when a condition is true

fn advance_n(&mut self, n: u32)

Advances the cursor’s position by n.

fn close(&mut self) -> TemporalResult<()>

Closes the current cursor by checking if all contents have been consumed. If not, returns an error for invalid syntax.