Crate boa_temporal

Boa’s boa_temporal crate is an engine agnostic implementation of ECMAScript’s Temporal.

IMPORTANT NOTE: Please note that this library is actively being developed and is very much experimental and NOT STABLE.

Temporal is the Stage 3 proposal for ECMAScript that provides new JS objects and functions for working with dates and times that fully supports time zones and non-gregorian calendars.

This library’s primary source is the Temporal Proposal specification.

About Boa

Boa is an open-source, experimental ECMAScript Engine written in Rust for lexing, parsing and executing ECMAScript/JavaScript. Currently, Boa supports some of the language. More information can be viewed at Boa’s website.

Try out the most recent release with Boa’s live demo playground.

Boa Crates

boa_ast - Boa’s ECMAScript Abstract Syntax Tree.

boa_engine - Boa’s implementation of ECMAScript builtin objects and execution.

boa_gc - Boa’s garbage collector.

boa_interner - Boa’s string interner.

boa_parser - Boa’s lexer and parser.

boa_profiler - Boa’s code profiler.

boa_icu_provider - Boa’s ICU4X data provider.

boa_runtime - Boa’s WebAPI features.

Modules

components

The primary date-time components provided by Temporal.

error

This module implements TemporalError.

fields

This module implements a native Rust TemporalField and components.

iso

This module implements the internal ISO field slots.

options

Native implementation of the Temporal options.

parser

This module implements parsing for ISO 8601 grammar.

Macros

assert_syntax

assert_syntax! is a parser specific utility macro for asserting a syntax test, and returning a SyntaxError with the provided message if the test fails.

Structs

TemporalError

The error type for boa_temporal.

TemporalFields

TemporalFields acts as a native Rust implementation of the fields object

TinyAsciiStr

Constants

MS_PER_DAY

Milliseconds per day constant: 8.64e+7

NS_PER_DAY

Nanoseconds per day constant: 8.64e+13

Type Aliases

TemporalResult

The Temporal result type