Trait boa_temporal::components::calendar::CalendarProtocol

pub trait CalendarProtocol: Clone {

    // Required methods
    fn date_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut dyn Any
    ) -> TemporalResult<Date<Self>>;
    fn year_month_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut dyn Any
    ) -> TemporalResult<YearMonth<Self>>;
    fn month_day_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut dyn Any
    ) -> TemporalResult<MonthDay<Self>>;
    fn date_add(
        &self,
        date: &Date<Self>,
        duration: &Duration,
        overflow: ArithmeticOverflow,
        context: &mut dyn Any
    ) -> TemporalResult<Date<Self>>;
    fn date_until(
        &self,
        one: &Date<Self>,
        two: &Date<Self>,
        largest_unit: TemporalUnit,
        context: &mut dyn Any
    ) -> TemporalResult<Duration>;
    fn era(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<Option<TinyAsciiStr<16>>>;
    fn era_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<Option<i32>>;
    fn year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<i32>;
    fn month(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u8>;
    fn month_code(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<TinyAsciiStr<4>>;
    fn day(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u8>;
    fn day_of_week(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn day_of_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn week_of_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn year_of_week(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<i32>;
    fn days_in_week(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn days_in_month(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn days_in_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn months_in_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<u16>;
    fn in_leap_year(
        &self,
        date_like: &CalendarDateLike<Self>,
        context: &mut dyn Any
    ) -> TemporalResult<bool>;
    fn fields(
        &self,
        fields: Vec<String>,
        context: &mut dyn Any
    ) -> TemporalResult<Vec<String>>;
    fn merge_fields(
        &self,
        fields: &TemporalFields,
        additional_fields: &TemporalFields,
        context: &mut dyn Any
    ) -> TemporalResult<TemporalFields>;
    fn identifier(&self, context: &mut dyn Any) -> TemporalResult<String>;

}

A trait for implementing a Builtin Calendar’s Calendar Protocol in Rust.

Required Methods

fn date_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut dyn Any ) -> TemporalResult<Date<Self>>

Creates a Temporal.PlainDate object from provided fields.

fn year_month_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut dyn Any ) -> TemporalResult<YearMonth<Self>>

Creates a Temporal.PlainYearMonth object from the provided fields.

fn month_day_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut dyn Any ) -> TemporalResult<MonthDay<Self>>

Creates a Temporal.PlainMonthDay object from the provided fields.

fn date_add( &self, date: &Date<Self>, duration: &Duration, overflow: ArithmeticOverflow, context: &mut dyn Any ) -> TemporalResult<Date<Self>>

Returns a Temporal.PlainDate based off an added date.

fn date_until( &self, one: &Date<Self>, two: &Date<Self>, largest_unit: TemporalUnit, context: &mut dyn Any ) -> TemporalResult<Duration>

Returns a Temporal.Duration representing the duration between two dates.

fn era( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<Option<TinyAsciiStr<16>>>

Returns the era for a given temporaldatelike.

fn era_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<Option<i32>>

Returns the era year for a given temporaldatelike

fn year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<i32>

Returns the year for a given temporaldatelike

fn month( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u8>

Returns the month for a given temporaldatelike

fn month_code( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<TinyAsciiStr<4>>

Returns the monthCode for a given temporaldatelike

fn day( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u8>

Returns the day for a given temporaldatelike

fn day_of_week( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns a value representing the day of the week for a date.

fn day_of_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns a value representing the day of the year for a given calendar.

fn week_of_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns a value representing the week of the year for a given calendar.

fn year_of_week( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<i32>

Returns the year of a given week.

fn days_in_week( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the days in a week for a given calendar.

fn days_in_month( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the days in a month for a given calendar.

fn days_in_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the days in a year for a given calendar.

fn months_in_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<u16>

Returns the months in a year for a given calendar.

fn in_leap_year( &self, date_like: &CalendarDateLike<Self>, context: &mut dyn Any ) -> TemporalResult<bool>

Returns whether a value is within a leap year according to the designated calendar.

fn fields( &self, fields: Vec<String>, context: &mut dyn Any ) -> TemporalResult<Vec<String>>

Return the fields for a value.

fn merge_fields( &self, fields: &TemporalFields, additional_fields: &TemporalFields, context: &mut dyn Any ) -> TemporalResult<TemporalFields>

Merge fields based on the calendar and provided values.

fn identifier(&self, context: &mut dyn Any) -> TemporalResult<String>

Debug name

Object Safety

This trait is not object safe.